requirements.txt
accelerate
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.11.0
asttokens==2.4.1
attrs==25.4.0
certifi==2025.10.5
charset_normalizer==3.4.4
comm==0.2.2
contourpy
cycler==0.12.1
datasets
debugpy==1.8.1
decorator==5.1.1
dill==0.4.0
executing==2.0.1
filelock==3.20.0
fonttools==4.53.0
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf_xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface_hub
idna==3.11
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jinja2==3.1.6
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==3.0.2
matplotlib
matplotlib_inline==0.1.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx  # Removed pin: Was explicitly incompatible (3.5 requires Python 3.11+)
nose==1.3.7
numpy
opencv_contrib_python
opencv_contrib_python_headless
opencv_python
opencv_python_headless
packaging==25.0
pandas
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
# Pillow_SIMD==9.5.0.post2
pip==25.2
platformdirs==3.9.1
prompt_toolkit==3.0.47
propcache==0.4.1
protobuf==6.33.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.2
pyarrow
pygments==2.18.0
pyparsing==3.1.2
python_dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
regex==2024.11.6
requests==2.32.5
safetensors==0.5.3
scipy
sentencepiece==0.2.0
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
stack_data==0.6.3
sympy==1.14.0
timm
tokenizers==0.22.1
torch
torchvision
tornado==6.3.3
tqdm==4.67.1
traitlets==5.14.3
transformers
typing_extensions==4.15.0
tzdata==2024.1
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
xxhash==3.5.0
yarl==1.22.0